requirements.txt
attrs==19.3.0
dataclasses==0.8.0
matplotlib==3.3.4
nest-asyncio==1.4.0
numpy==1.18.5
scikit-learn==0.24.2
scipy==1.4.1
six==1.15.0
tensorflow==2.2.0
tensorflow-federated==0.16.1
